fix: wrap pre-screen useSearchParams in Suspense boundary
